Decompose harness/src/lib.rs into focused modules (issue #65)
Split the 3500-line lib.rs into loop_.rs, hooks.rs, context.rs
(not created—see notes), prompt.rs, and added
crates/harness/harness.spec.md.
