+++ crates/harness/src/lib.rs
@@ -1,14 +1,35 @@
-use agents::{AgentHooks, HookCommand, HookEntry};
-use anthropic::{AnthropicClient, MessageRequest, MessageResponse};
+mod hooks;
+mod loop_;
+mod prompt;
+
 #[cfg(test)]
 use async_trait::async_trait;
+
+pub use loop_::run;
+pub use loop_::resolve_session_cwd;
+pub use anthropic::StubClient;
+
+#[cfg(test)]
+use anthropic::{AnthropicClient, MessageRequest, MessageResponse};
+#[cfg(test)]
 use chrono::Utc;
-use regex::Regex;
-use std::path::{Path, PathBuf};
-use std::sync::Arc;
-use tokio::sync::{broadcast, mpsc};
-use types::{ContentBlock, ContentBlockDelta, Role, Session, SessionEvent, SessionStatus, Turn};
+#[cfg(test)]
+use types::{ContentBlock, Role, SessionEvent};
+#[cfg(test)]
 use uuid::Uuid;
+#[cfg(test)]
+use loop_::{
+    is_rate_limit, load_history, max_retries, persist_user_message,
+    resolve_session_cwd_with_root, run_tool_dispatch_loop,
+};
+#[cfg(test)]
+use prompt::build_system_prompt;
+#[cfg(test)]
+use hooks::{matching_hook_entries, run_hook, run_post_tool_use_hooks, run_pre_tool_use_hooks, run_stop_hooks};
+
+use std::path::PathBuf;
+use std::sync::Arc;
+use types::Session;
 
 #[derive(Debug, thiserror::Error)]
 pub enum Error {
@@ -20,310 +41,6 @@ pub enum Error {
 
 pub type Result<T> = std::result::Result<T, Error>;
 
-// ── System prompt construction ────────────────────────────────────────────────
-
-/// Build the preamble block that is prepended to every agent system prompt when
-/// the git root is known. Returns `None` when the root cannot be determined so
-/// callers can omit the preamble without failing.
-pub(crate) fn build_preamble(root: &Path) -> String {
-    let repo_name = root.file_name().unwrap_or_default().to_string_lossy();
-    format!(
-        "You are running in the ns2 agent harness.\nWorking directory / git root: {}\nRepository: {}\n",
-        root.display(),
-        repo_name,
-    )
-}
-
-/// Assemble the final system prompt from an optional git root, an optional
-/// agent definition directory, the session agent name, and the effective root
-/// used for project config loading.
-///
-/// Returns `None` when there is nothing useful to send (no preamble, no agent body).
-pub(crate) fn build_system_prompt(
-    effective_root: Option<&Path>,
-    agents_dir: Option<&Path>,
-    agent_name: Option<&str>,
-) -> Option<String> {
-    // Build preamble if we know where we are.
-    let preamble: Option<String> = effective_root.map(build_preamble);
-
-    // Load agent body (+ optional project config).
-    let agent_body_and_project: Option<String> = agent_name.and_then(|name| {
-        let dir = agents_dir?;
-        agents::load_agent(dir, name)
-    }).and_then(|def| {
-        let agent_body = def.body;
-        if def.include_project_config {
-            let project = effective_root
-                .and_then(agents::load_project_config)
-                .unwrap_or_default();
-            if project.is_empty() {
-                if agent_body.is_empty() { None } else { Some(agent_body) }
-            } else {
-                Some(format!("{agent_body}\n\n{project}"))
-            }
-        } else {
-            if agent_body.is_empty() { None } else { Some(agent_body) }
-        }
-    });
-
-    match (preamble, agent_body_and_project) {
-        (Some(pre), Some(body)) => Some(format!("{pre}{body}")),
-        (_,         None)       => None,   // no agent body → no system prompt, even with a preamble
-        (None,      body)       => body,
-    }
-}
-
-// ── Hook execution ────────────────────────────────────────────────────────────
-
-/// Run a single hook command with JSON on stdin.
-/// Returns `(exit_code, stdout, stderr)`.
-/// Kills the process and returns exit_code=1 on timeout.
-async fn run_hook(cmd: &HookCommand, stdin_json: &str) -> (i32, String, String) {
-    use tokio::io::{AsyncReadExt, AsyncWriteExt};
-    use tokio::process::Command;
-    use tokio::time::{sleep, Duration};
-
-    let mut child = match Command::new("sh")
-        .arg("-c")
-        .arg(&cmd.command)
-        .stdin(std::process::Stdio::piped())
-        .stdout(std::process::Stdio::piped())
-        .stderr(std::process::Stdio::piped())
-        .spawn()
-    {
-        Ok(c) => c,
-        Err(e) => {
-            tracing::warn!("failed to spawn hook command '{}': {e}", cmd.command);
-            return (1, String::new(), format!("failed to spawn: {e}"));
-        }
-    };
-
-    // Write stdin and close the pipe
-    if let Some(mut stdin) = child.stdin.take() {
-        let _ = stdin.write_all(stdin_json.as_bytes()).await;
-    }
-
-    // Take stdout/stderr handles for reading
-    let mut stdout_handle = child.stdout.take();
-    let mut stderr_handle = child.stderr.take();
-
-    // Read stdout and stderr concurrently with a timeout on the entire operation
-    let duration = Duration::from_secs(cmd.timeout);
-    tokio::select! {
-        result = async {
-            let mut stdout_buf = Vec::new();
-            let mut stderr_buf = Vec::new();
-            if let Some(ref mut h) = stdout_handle {
-                let _ = h.read_to_end(&mut stdout_buf).await;
-            }
-            if let Some(ref mut h) = stderr_handle {
-                let _ = h.read_to_end(&mut stderr_buf).await;
-            }
-            let status = child.wait().await;
-            (status, stdout_buf, stderr_buf)
-        } => {
-            let (status, stdout_buf, stderr_buf) = result;
-            let exit_code = status.ok().and_then(|s| s.code()).unwrap_or(1);
-            let stdout = String::from_utf8_lossy(&stdout_buf).into_owned();
-            let stderr = String::from_utf8_lossy(&stderr_buf).into_owned();
-            (exit_code, stdout, stderr)
-        }
-        _ = sleep(duration) => {
-            tracing::warn!("hook command '{}' timed out after {}s", cmd.command, cmd.timeout);
-            let _ = child.kill().await;
-            (1, String::new(), format!("hook timed out after {}s", cmd.timeout))
-        }
-    }
-}
-
-/// Find all `HookEntry`s whose matcher regex matches `tool_name`.
-fn matching_hook_entries<'a>(entries: &'a [HookEntry], tool_name: &str) -> Vec<&'a HookEntry> {
-    entries
-        .iter()
-        .filter(|e| {
-            e.matcher
-                .as_deref()
-                .map(|pat| Regex::new(pat).map(|re| re.is_match(tool_name)).unwrap_or(false))
-                .unwrap_or(false)
-        })
-        .collect()
-}
-
-/// Run PreToolUse hooks for `tool_name`.
-/// Returns `Some(blocked_message)` if any hook exits non-zero (tool should be skipped),
-/// or `None` if all hooks pass.
-async fn run_pre_tool_use_hooks(
-    hooks: &AgentHooks,
-    tool_name: &str,
-    tool_input: &serde_json::Value,
-) -> Option<String> {
-    let stdin = serde_json::json!({
-        "tool_name": tool_name,
-        "tool_input": tool_input,
-    });
-    let stdin_str = stdin.to_string();
-
-    for entry in matching_hook_entries(&hooks.pre_tool_use, tool_name) {
-        for cmd in &entry.hooks {
-            let (exit_code, _stdout, stderr) = run_hook(cmd, &stdin_str).await;
-            if exit_code != 0 {
-                return Some(if stderr.is_empty() {
-                    format!("Hook blocked tool '{tool_name}' (exit {exit_code})")
-                } else {
-                    stderr
-                });
-            }
-        }
-    }
-    None
-}
-
-/// Run PostToolUse hooks for `tool_name`. Exit code is always ignored.
-async fn run_post_tool_use_hooks(
-    hooks: &AgentHooks,
-    tool_name: &str,
-    tool_input: &serde_json::Value,
-    tool_result: &str,
-) {
-    let stdin = serde_json::json!({
-        "tool_name": tool_name,
-        "tool_input": tool_input,
-        "tool_result": tool_result,
-    });
-    let stdin_str = stdin.to_string();
-
-    for entry in matching_hook_entries(&hooks.post_tool_use, tool_name) {
-        for cmd in &entry.hooks {
-            run_hook(cmd, &stdin_str).await;
-        }
-    }
-}
-
-/// Run Stop hooks.
-/// Returns `Some(injected_message)` if any hook exits non-zero (the message should be
-/// injected into the conversation to continue the loop), or `None` to allow completion.
-///
-/// Exit code 127 (command not found) is treated as a configuration error and fails
-/// open — the hook is skipped and the session is allowed to complete normally.
-/// This prevents an infinite loop when a hook script cannot be found (e.g. because
-/// `$CLAUDE_PROJECT_DIR` is unset and the path cannot be resolved).
-async fn run_stop_hooks(hooks: &AgentHooks, session_id: Uuid) -> Option<String> {
-    let stdin = serde_json::json!({ "session_id": session_id.to_string() });
-    let stdin_str = stdin.to_string();
-
-    for entry in &hooks.stop {
-        for cmd in &entry.hooks {
-            let (exit_code, stdout, _stderr) = run_hook(cmd, &stdin_str).await;
-            if exit_code != 0 {
-                // Exit 127 = command not found: configuration error, not a deliberate
-                // block. Fail open so the agent is not trapped in an infinite loop.
-                if exit_code == 127 {
-                    tracing::warn!(
-                        "stop hook command not found (exit 127): '{}' — failing open",
-                        cmd.command
-                    );
-                    continue;
-                }
-                return Some(if stdout.is_empty() {
-                    format!("Stop hook exited with code {exit_code}")
-                } else {
-                    stdout
-                });
-            }
-        }
-    }
-    None
-}
-
-// ── 429 retry logic ──────────────────────────────────────────────────────────
-
-/// Read the max-retry count from `NS2_MAX_RETRIES` (default 5).
-fn max_retries() -> u32 {
-    std::env::var("NS2_MAX_RETRIES")
-        .ok()
-        .and_then(|v| v.parse::<u32>().ok())
-        .unwrap_or(5)
-}
-
-/// Returns `true` if the error is an Anthropic 429 rate-limit response.
-fn is_rate_limit(err: &anthropic::Error) -> bool {
-    matches!(err, anthropic::Error::Api { status: 429, .. })
-}
-
-/// Call `client.complete(request)` with exponential-backoff retry on 429.
-///
-/// - Retries up to `NS2_MAX_RETRIES` times (default 5).
-/// - Initial delay: 10 s, doubling each retry, capped at 120 s.
-/// - Non-429 errors propagate immediately without retrying.
-///
-/// In tests using `#[tokio::test(start_paused = true)]`, `tokio::time::sleep`
-/// returns instantly, so no real wall-clock time is consumed.
-async fn complete_with_retry(
-    client: &Arc<dyn AnthropicClient>,
-    request: MessageRequest,
-) -> anthropic::Result<MessageResponse> {
-    use tokio::time::{sleep, Duration};
-
-    let retries = max_retries();
-    let mut delay_ms: u64 = 10_000;
-    const MAX_DELAY_MS: u64 = 120_000;
-
-    let mut attempt = 0u32;
-    loop {
-        match client.complete(request.clone()).await {
-            Ok(resp) => return Ok(resp),
-            Err(err) if is_rate_limit(&err) && attempt < retries => {
-                attempt += 1;
-                tracing::warn!(
-                    attempt,
-                    retries,
-                    delay_ms,
-                    "Anthropic 429 rate-limit; retrying after {delay_ms}ms"
-                );
-                sleep(Duration::from_millis(delay_ms)).await;
-                delay_ms = (delay_ms * 2).min(MAX_DELAY_MS);
-            }
-            Err(err) => return Err(err),
-        }
-    }
-}
-
-// ── Worktree management ───────────────────────────────────────────────────────
-
-/// Resolve the session's working directory based on its associated issue.
-///
-/// - If the session has an associated issue with a non-empty `branch`:
-///   reads `ns2.toml`, computes `<base>/<branch>`, ensures the worktree
-///   exists, and returns the worktree path.
-/// - Otherwise: returns `None` (use git root as cwd).
-pub async fn resolve_session_cwd(
-    db: &Arc<dyn db::Db>,
-    session_id: Uuid,
-) -> Option<PathBuf> {
-    resolve_session_cwd_with_root(db, session_id, workspace::git_root().await).await
-}
-
-/// Inner implementation that accepts an explicit `git_root` — injectable for tests.
-pub(crate) async fn resolve_session_cwd_with_root(
-    db: &Arc<dyn db::Db>,
-    session_id: Uuid,
-    git_root: Option<PathBuf>,
-) -> Option<PathBuf> {
-    let issues = db.list_issues_by_session_id(session_id).await.ok()?;
-    let branch = issues.into_iter().find_map(|i| {
-        if i.branch.is_empty() { None } else { Some(i.branch) }
-    })?;
-
-    let git_root = git_root?;
-    let config = workspace::read_ns2_config(&git_root);
-    let worktree_path = config.worktree_base.join(&branch);
-
-    workspace::ensure_worktree(&git_root, &worktree_path, &branch).await
-}
-
-pub use anthropic::StubClient;
-
 pub struct HarnessConfig {
     pub session: Session,
     pub model: String,
@@ -332,350 +49,6 @@ pub struct HarnessConfig {
     pub git_root: Option<PathBuf>,
 }
 
-/// Load conversation history from the DB for a session.
-/// Returns turns in order, each as `(Role, Vec<ContentBlock>)`.
-/// Turns with mixed roles are grouped by the role stored on each block;
-/// consecutive blocks with the same role are merged into one entry.
-async fn load_history(
-    db: &Arc<dyn db::Db>,
-    session_id: Uuid,
-) -> Result<Vec<(Role, Vec<ContentBlock>)>> {
-    let turns = db.list_turns(session_id).await?;
-    let mut history: Vec<(Role, Vec<ContentBlock>)> = Vec::new();
-
-    for turn in &turns {
-        let blocks = db.list_content_blocks(turn.id).await?;
-        if blocks.is_empty() {
-            continue;
-        }
-        // Each turn is stored with a consistent role; group all blocks under one entry.
-        // If blocks have mixed roles (shouldn't happen in practice), group by first role.
-        let role = blocks[0].0.clone();
-        let content: Vec<ContentBlock> = blocks.into_iter().map(|(_, b)| b).collect();
-
-        // Merge with previous entry if same role
-        if let Some(last) = history.last_mut() {
-            if last.0 == role {
-                last.1.extend(content);
-                continue;
-            }
-        }
-        history.push((role, content));
-    }
-
-    Ok(history)
-}
-
-/// Persist a user message as a turn+block in the DB and emit events.
-async fn persist_user_message(
-    db: &Arc<dyn db::Db>,
-    session_id: Uuid,
-    message: &str,
-    event_tx: &broadcast::Sender<SessionEvent>,
-) -> Result<Turn> {
-    let user_turn = Turn {
-        id: Uuid::new_v4(),
-        session_id,
-        token_count: None,
-        created_at: Utc::now(),
-    };
-    db.create_turn(&user_turn).await?;
-    let user_block = ContentBlock::Text { text: message.to_string() };
-    db.create_content_block(user_turn.id, 0, &Role::User, &user_block).await?;
-    let _ = event_tx.send(SessionEvent::TurnStarted { turn: user_turn.clone() });
-    let _ = event_tx.send(SessionEvent::ContentBlockDelta {
-        turn_id: user_turn.id,
-        index: 0,
-        delta: ContentBlockDelta::TextDelta { text: message.to_string() },
-    });
-    let _ = event_tx.send(SessionEvent::ContentBlockDone {
-        turn_id: user_turn.id,
-        index: 0,
-        block: user_block,
-    });
-    let _ = event_tx.send(SessionEvent::TurnDone { turn_id: user_turn.id });
-    Ok(user_turn)
-}
-
-/// Run the tool dispatch loop for a single LLM turn sequence.
-/// `messages` should already contain the full history including the new user message.
-async fn run_tool_dispatch_loop(
-    config: &HarnessConfig,
-    client: &Arc<dyn AnthropicClient>,
-    db: &Arc<dyn db::Db>,
-    event_tx: &broadcast::Sender<SessionEvent>,
-    hooks: &AgentHooks,
-    system: Option<String>,
-    mut messages: Vec<(Role, Vec<ContentBlock>)>,
-) -> Result<Option<String>> {
-    let tool_definitions: Vec<types::ToolDefinition> =
-        config.tools.iter().map(|t| t.definition()).collect();
-
-    loop {
-        let request = MessageRequest {
-            model: config.model.clone(),
-            system: system.clone(),
-            messages: messages.clone(),
-            max_tokens: 64000,
-            tools: tool_definitions.clone(),
-        };
-
-        let response = match complete_with_retry(client, request).await {
-            Ok(r) => r,
-            Err(err) => {
-                // Retries exhausted (429) or non-retryable error → emit error event
-                // and break out of the dispatch loop. Non-429 errors are re-raised
-                // to the caller so the harness task can fail clearly.
-                if is_rate_limit(&err) {
-                    let _ = event_tx.send(SessionEvent::Error {
-                        message: format!(
-                            "rate limited after {retries} retries: {err}",
-                            retries = max_retries()
-                        ),
-                    });
-                    break;
-                }
-                return Err(Error::Anthropic(err));
-            }
-        };
-
-        // Create assistant turn in DB
-        let turn = Turn {
-            id: Uuid::new_v4(),
-            session_id: config.session.id,
-            token_count: Some((response.input_tokens + response.output_tokens) as i64),
-            created_at: Utc::now(),
-        };
-        db.create_turn(&turn).await?;
-        let _ = event_tx.send(SessionEvent::TurnStarted { turn: turn.clone() });
-
-        // Store and emit content blocks
-        for (index, block) in response.content.iter().enumerate() {
-            let index = index as u32;
-            if let ContentBlock::Text { text } = block {
-                let _ = event_tx.send(SessionEvent::ContentBlockDelta {
-                    turn_id: turn.id,
-                    index,
-                    delta: ContentBlockDelta::TextDelta { text: text.clone() },
-                });
-            }
-            db.create_content_block(turn.id, index as i64, &Role::Assistant, block).await?;
-            let _ = event_tx.send(SessionEvent::ContentBlockDone {
-                turn_id: turn.id,
-                index,
-                block: block.clone(),
-            });
-        }
-
-        let _ = event_tx.send(SessionEvent::TurnDone { turn_id: turn.id });
-
-        match response.stop_reason.as_str() {
-            "tool_use" => { /* dispatch tools below */ }
-            "end_turn" => {
-                // Run Stop hooks; if any exit non-zero, return their stdout as an injected message
-                if let Some(injected) = run_stop_hooks(hooks, config.session.id).await {
-                    return Ok(Some(injected));
-                }
-                break;
-            }
-            "max_tokens" => {
-                let _ = event_tx.send(SessionEvent::Error {
-                    message: "session hit max_tokens limit".to_string(),
-                });
-                break;
-            }
-            other => {
-                tracing::warn!("unknown stop_reason: {other}");
-                break;
-            }
-        }
-
-        // Add assistant turn to history
-        messages.push((Role::Assistant, response.content.clone()));
-
-        // Execute tool calls and build tool result turn
-        let mut tool_result_blocks: Vec<ContentBlock> = Vec::new();
-
-        for block in &response.content {
-            if let ContentBlock::ToolUse { id, name, input } = block {
-                // Run PreToolUse hooks
-                let result = if let Some(blocked) =
-                    run_pre_tool_use_hooks(hooks, name, input).await
-                {
-                    // Hook blocked the tool — return hook stderr as the tool result
-                    blocked
-                } else {
-                    // Run the actual tool
-                    let tool_output = match config.tools.iter().find(|t| t.definition().name == *name) {
-                        Some(tool) => match tool.execute(input.clone()).await {
-                            Ok(output) => output,
-                            Err(e) => format!("Error: {e}"),
-                        },
-                        None => format!("Error: unknown tool '{name}'"),
-                    };
-                    // Run PostToolUse hooks (exit code ignored)
-                    run_post_tool_use_hooks(hooks, name, input, &tool_output).await;
-                    tool_output
-                };
-                tool_result_blocks.push(ContentBlock::ToolResult {
-                    tool_use_id: id.clone(),
-                    content: result,
-                });
-            }
-        }
-
-        // Store tool result turn in DB
-        let tool_result_turn = Turn {
-            id: Uuid::new_v4(),
-            session_id: config.session.id,
-            token_count: None,
-            created_at: Utc::now(),
-        };
-        db.create_turn(&tool_result_turn).await?;
-        let _ = event_tx.send(SessionEvent::TurnStarted { turn: tool_result_turn.clone() });
-
-        for (index, block) in tool_result_blocks.iter().enumerate() {
-            let index = index as u32;
-            db.create_content_block(tool_result_turn.id, index as i64, &Role::User, block)
-                .await?;
-            let _ = event_tx.send(SessionEvent::ContentBlockDone {
-                turn_id: tool_result_turn.id,
-                index,
-                block: block.clone(),
-            });
-        }
-
-        let _ = event_tx.send(SessionEvent::TurnDone { turn_id: tool_result_turn.id });
-
-        // Add tool result turn to history and loop
-        messages.push((Role::User, tool_result_blocks));
-    }
-
-    Ok(None)
-}
-
-pub async fn run(
-    mut config: HarnessConfig,
-    client: Arc<dyn AnthropicClient>,
-    db: Arc<dyn db::Db>,
-    event_tx: broadcast::Sender<SessionEvent>,
-    mut msg_rx: mpsc::Receiver<String>,
-) -> Result<()> {
-    // Resolve the session's working directory once at startup.
-    // If the session has an associated issue with a non-empty branch, create/reuse
-    // a git worktree and set cwd to the worktree path.
-    let session_cwd = resolve_session_cwd(&db, config.session.id).await;
-
-    // If we resolved a cwd, rebuild the standard tool set with that cwd so all
-    // file operations and shell commands run relative to the worktree.
-    if let Some(ref cwd) = session_cwd {
-        config.tools = vec![
-            Arc::new(tools::BashTool { cwd: Some(cwd.clone()) }),
-            Arc::new(tools::ReadTool { cwd: Some(cwd.clone()) }),
-            Arc::new(tools::WriteTool { cwd: Some(cwd.clone()) }),
-            Arc::new(tools::EditTool { cwd: Some(cwd.clone()) }),
-        ];
-    }
-
-    // Resolve the effective git root (injected in tests; discovered via git in production).
-    let effective_root = if let Some(root) = config.git_root.clone() {
-        Some(root)
-    } else {
-        workspace::git_root().await
-    };
-    let agents_dir = effective_root.as_ref().map(|r| r.join(".ns2").join("agents"));
-
-    // Build a preamble that tells the agent where it is running.
-    // If the git root cannot be determined, the preamble is omitted (no failure).
-    // Pre-compute the system prompt once (it does not change across turns).
-    // The preamble (if available) is always prepended before the agent body so
-    // every agent session knows its working directory and repository name.
-    let system: Option<String> = build_system_prompt(
-        effective_root.as_deref(),
-        agents_dir.as_deref(),
-        config.session.agent.as_deref(),
-    );
-
-    // Load hooks from the agent definition (once, at harness start).
-    // When include_project_config is true, also load project hooks from
-    // .claude/settings.json and merge them (agent hooks take precedence).
-    let hooks = {
-        let agent_def = config.session.agent.as_deref().and_then(|name| {
-            let dir = agents_dir.as_ref()?;
-            agents::load_agent(dir, name)
-        });
-
-        match agent_def {
-            None => AgentHooks::default(),
-            Some(def) => {
-                let agent_hooks = def.hooks;
-                if def.include_project_config {
-                    let project_hooks = effective_root.as_deref()
-                        .map(agents::load_project_hooks)
-                        .unwrap_or_default();
-                    agents::merge_hooks(agent_hooks, project_hooks)
-                } else {
-                    agent_hooks
-                }
-            }
-        }
-    };
-
-    loop {
-        // Wait for the next user message. When the sender is dropped, recv() returns None → exit.
-        let message = match msg_rx.recv().await {
-            Some(m) => m,
-            None => break,
-        };
-
-        // Transition session to Running
-        db.update_session_status(config.session.id, SessionStatus::Running).await?;
-
-        // Persist user message and emit events
-        persist_user_message(&db, config.session.id, &message, &event_tx).await?;
-
-        // Load full conversation history from DB (including the message we just stored)
-        let history = load_history(&db, config.session.id).await?;
-
-        // Run the tool dispatch loop. It returns Some(injected_message) when a Stop
-        // hook blocks completion and injects a message for the next turn.
-        let mut current_history = history;
-        loop {
-            match run_tool_dispatch_loop(
-                &config,
-                &client,
-                &db,
-                &event_tx,
-                &hooks,
-                system.clone(),
-                current_history,
-            )
-            .await?
-            {
-                None => break, // normal completion
-                Some(injected) => {
-                    // Stop hook rejected; inject the message and continue
-                    persist_user_message(
-                        &db,
-                        config.session.id,
-                        &injected,
-                        &event_tx,
-                    )
-                    .await?;
-                    current_history = load_history(&db, config.session.id).await?;
-                }
-            }
-        }
-
-        // Mark session completed and emit done event
-        db.update_session_status(config.session.id, SessionStatus::Completed).await?;
-        let _ = event_tx.send(SessionEvent::SessionDone { session_id: config.session.id });
-
-        // Loop back to wait for next message
-    }
-
-    Ok(())
-}
 
 #[cfg(test)]
 mod tests {
